Anti-Rollback Support: Secure Version passed from HBBL to HDAT
This commit adds the support for the Hostboot bootloader (HBBL)
to share what the Minimum FW Secure Version level was built-in
to the master processor's SBE.  It is then used throughout the hostboot
stack via the Bootloader-to-Hostboot data interface. Ultimately,
the value gets placed into HDAT.

Please note that the secure versions encapsulated in the headers of
PNOR partitions are not called the Minimum FW Secure Version as
they are actually tested against the system's Minimum FW Secure Version.

Change-Id: Ie504dcef40abed66c4b99ef9ecc4346e5acda32d
RTC:254577
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99965
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>