Add Denali specific simics commands to autopatch and boot FSP
Adds code that will validate an environment variable HB_NFS_DIR if it
exists. Should that variable exist and be valid then it will define the
simics environment variable $hb_nfsdir to be the validated path. It will
trigger the fsp_autoboot.simics script to run which will autopatch and
autoboot Denali simics when FSP standby is reached.

Change-Id: I7fbd5588b6cca819848d7ce81cc28f3e2e6b593c
RTC:246824
RTC:246815
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99515
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>