scominfo: Move CLKADJ logic to core scope
Even though the CLKADJ logic is on the EQ's perv SCOM loop each core's
individual adjust logic is attached to that core's own perv loop instance.

Update the scominfo code so that the clkadj satellite ID is detected as
a core address whereas all other satellites on the perv loop are still
EQ scope.

Change-Id: Iad2e9d45f8fef68fc3120034ed684baf3c8d61cc
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99524
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Brent Wieman <bwieman@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99531
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>