Set OCC PDR values based on PDR manager values instead of attributes
Previously we were relying on the ATTR_PLDM_ENTITY_ID_INFO being set
on the PROC target prior to the OCC pdrs being created. This was not
happening and caused the values to be incorrect. To prevent this
we will instead provide the entity type, entity instance, and container
id of the PDR by looking up the value of the PROC's entity PDR.

Change-Id: I36850df1b95eb1e8f607979ef53294b2afc172a3
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99528
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>