Fix bug in PLDM state set enumeration value encoding
Before this commit, the PLDM code was encoding state set enumeration
values into bitsets by translating the value "1" to position 0 in the
bitset. It should translate the value "0" to position 0, 1 to 1, etc.

Also fix a bug where the effecter state in the SetStateEffecterStates
command handler was being compared against the bitset for the
enumeration value, rather than the enumeration value itself.

Change-Id: Ib1d0dff612e8175956918c561ef04645ebd5e64f
RTC: 250657
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99534
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>