Do not start polling KCS status register until MCTP bus is intitialized
There is a bug that leaves us exposed to a potential program crash. We
start polling the status reg before the bus is initialized. If we happen
to see a MCTP message during this time we will attempt to handle it and
will assert because the bus is not set up yet.

Change-Id: Ib5b21ad56f5b8bd8bbaa9956501bb3e6ca1dd321
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99535
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>