vbu/contained: Improve thread seed generation
Introduce some improvements to the exerciser thread seed generation
during cache- and chip-contained RUNN IPLs. First, move the seed
generation to `proc_contained_run`, ensure generated seeds are always
consistent on a per-core basis, and finally, redefine the enums for
ATTR_RUNN_CORE_SEED_SELECT to correspond to core numbers. More details
on the first two below.

To support reproducible exerciser runs in a lab environment the core0
scanrings are saved after the `proc_contained_ipl` step. This saved,
"golden", scanring state can be loaded on a subsequent IPL to reproduce
the same inits. Previously, the thread seeds were randomized during the
`proc_contained_ipl` step and would thus be included in the "golden"
rings state. This is not ideal since this means that core0's thread
seeds would always be used for re-runs using this saved scanring state.

Instead, randomize and apply the exerciser thread seeds during
`proc_contained_run`.

Previously the seeds were generated in a loop over the active cores.
This is problematic and does not guarantee reproducible seeds since
not all cores in the system may be active/functional. We can't just
loop over cores and expect consistent results in all cases.

As an example, in a system with cores 0,1, and 6, core6 would be
seeded using the RNG state of core2. So instead, we re-seed the RNG
on every iteration, then advance the internal RNG state to the core
we need to generate seeds for. This guarantees that, given a
MASTER_SEED, we always generate the same unique sequence of random
seeds for every core irrespective of the system configuration.

Change-Id: I0a8b2ae161cab224e90278eea85e200a81fc5df4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99639
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99655
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>