Handle fapi2 getChildren<TARGET_TYPE_GENERICI2CSLAVE> for OCMB targets
There was a change in the HWP pmic_enable which calls getChildren on
an OCMB target for TARGET_TYPE_GENERICI2CSLAVE. Hostboot does not
support this target type so we were asserting when we failed to translate
it. The ideal behavior is to instead return a empty list of targets
so as a workaround that is what we will do.

Change-Id: Ie71c14672a5f92d6bf90e71cd25c8d96c7405ace
CQ: SW497075
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/99915
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>