Add dummy makefile for FSP SETUP pass so simics.tar is extracted
The makefile for src/hbfw/simics was removed in a previous commit. The
thought was that it was no longer needed and therefor could be removed.
However, in order to trigger the simics.tar file to be extracted and
allow FSP SETUP pass to complete, it would appear that a dummy file is
still necessary.

Change-Id: Iaf8b6166d633ca0b28e4b6f4577e96baa8d8c754
CQ:SW497437
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100040
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>