Anti-Rollback Support: Properly use "Minimum" Secure Version
This commit attempts to rename all variables and functions related to
the "Minimum" FW Secure Version that gets passed from the hostboot
bootloader (HBBL) to the hostboot instance.

Please note that secure versions encapsulated in the headers of
PNOR partitions do not receive this distiniction as they are tested
against the Minimum FW Secure Version.

Change-Id: I7fd8caa87a22ccdd583dac5c073e142ce0352bc0
RTC:254577
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100138
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>