Fix PRD's compile issue: Missed some items from previous commit
Dropped the '#ifdef __HOSTBOOT_MODULE' from the actual function
getAttrAsStdArr.  I removed it from the declaration itself but not
the defition.  This remedies that.

Replaced the '#ifdef __HOSTBOOT_MODULE' to '#if __cplusplus >= 201103L'
for methods tryGetAttr, trySetAttr, and setAttrFromStdArr.

In method mfgFlagAccessors.C::getMfgFlagCellIndex, changed assert
to TARG_ASSERT.

Change-Id: I6ad4043ee1890994093a6e80b3d7f2073442d7ef
CQ: SW496393
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100219
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>