MPIPL: Remove old P9 code path we used to power down secondary cores
In P9 the SBE only powered down the primary quad leaving the remaining
quads to be handled by Hostboot prior to issuing the continueMpipl
command to the seconday SBEs. In P10, the cores will be powered down
by the SBE as part of the continueMpipl flow so Hostboot does not need
to worry about this.

Change-Id: I3946a50f88f1619af0f398752e12c073a24bd599
RTC: 208841
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100222
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>