Key Clear: Fix Hostboot GPIO processing on multi-node systems
This commit fixes a bug where hostboot code was not properly setting
up ATTR_GPIO_INFO_PHYS_PRES for fleetwood systems which can have
different primary processors on multiple nodes.  Now, each processor
is given a default setting that works when it is the active primary
processor on that node.

Change-Id: I75923dda95f8185464696f295a594fa2d5e71510
CQ:SW495867
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100311
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>