Skip PDR exchange on MPIPLs until we resolve current issues
Currently the BMC is not removing the PDRs it gets from the Host
on an MPIPL trigger. When hostboot comes back up on the MPIPL we
are attempting to send the BMC our PDRs but the BMC already has
them so they do not send the PDR repo changed event that completes
the PDR exchange flow thus Hostboot is hanging. Until this is resolved
skip the PDR exchange and instead just get the copy of PDRs from the
BMC.

Change-Id: I21dfd006109b64f30be02e08d1d3f7416c669dab
RTC: 208841
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100353
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>