Allow case in MPIPL where ATTR_HOMER_PHYS_ADDR is 0
Previously during the PM_RESET we do in istep15.1 on the MPIPL if
the ATTR_HOMER_PHYS_ADDR is 0 for whatever reason, we would log
an error and fail to boot. In reality we could keep making progress
even if this value is 0. This commit updates the PM_RESET logic
to be tolerant of us not being able to map the HOMER region.

Change-Id: I160458453725219872f1851358ba7e7ccccde0a9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/100369
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>