Add attribute to note service processor's presence detection
Add FOUND_PRESENT_BY_SP to track the presence detection of parts
from the service processor.  This will be used to cross-check
with Hostboot's detection.

The attribute defaults to indicate that the service processor
did not attempt any detection so no cross-check is required.
This allows the code to work in non-FSP environments until
we need checking there as well.

Change-Id: Iafa4591300759ac80fb6574c7b5a27afeaca37eb
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101061
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>