Move temporary memory mirroring disablement to host_mss_attr_cleanup
Recently new code was added to host_mss_attr_cleanup to copy the memory
mirroring firmware policy into the HWP mirroring policy.  This bypassed
code in istep 14 that temporarily disables mirroring until Hostboot can
provide support for p10.  This commit moves that temporary disable from
istep 14 to host_mss_attr_cleanup so that the policy is consistent for the
duration of the boot.

Change-Id: I89f363454510e35f9da6041ece483198032237d9
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101073
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>