Fix PT entry offset calculations for the PLDM FRU->IPZ translation code
For the record size that starts every IPZ record that size does not include
the size itself, nor the large/small resource that mark the start and end
of the record. We were using this size to calculate the offsets to the
other records that get set in the pt entries. This commit updates the
logic to use set correct value of the VTOC record size, and also the correct
value for the PT entry offset fields.

Change-Id: I04a5751f37296fdd0fa5369e94cd7c545652d93f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101155
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Luis P Fernandez <luis.fernandez@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>