Reserve default SRC (BC801B99) to be used for TI discovery issues
There are some cases where the SP won't be able to read a valid
SRC from Hostboot's TI area.  For example, Hostboot could have
died horribly and been unable to modify memory, or there could
be issues reading mainstore itself.  This SRC will be used by
the attention handling code as a placeholder for any case where
a real SRC is not available.

Change-Id: I7d66efbeb8aaf2abb6aafcfbbfdef55c2dee82bd
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101327
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Brian J Stegmiller <bjs@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>