Force OMI Frequency to match what we booted with
This is a workaround until we can get the SPD in Simics to be in
sync with the MRW value.  When the two values are not in sync we
will attempt a SBE update to get things matched up.  That is fine
in product mode but makes every initial Simics boot fail.

Change-Id: Ib8aaae2e684b911e43c4c916a94204c20b620822
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101579
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>