Remove MPIPL references in exit cache contained HWP wrapper
Back in the early days of P9 MPIPL we thought we would have to do
some work in this istep during MPIPL but we later discovered we were
able to expand out of cache in thread activate earlier in istep 6. This
commit removes all of the left over logic in this istep wrapper that dealt
with MPIPL that never got cleaned up correctly.

RTC: 208841
Change-Id: I97718d3fab8485247f8c73e254f53b7ef96f5b18
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101604
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Luis P Fernandez <luis.fernandez@ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>