p10_proc_gettracearray: Remove PEC workaround
The workaround code was erroneously directing all SCOM accesses to
PCI1 even if it was asked to operate on PCI0. The workaround seems
to no longer be necessary (at least not on real HW) so remove it.
If that makes it crash in sim again, the root cause of that should
be investigated and fixed.

Change-Id: I0dfb8c0511a103d8c11fc9128e01694cb0b448f5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/101675
Reviewed-by: Nicholas Landi <nlandi@ibm.com>
Reviewed-by: Ricardo Mata <ricmata@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Dev-Ready: Joachim Fenkes <fenkes@de.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/106221
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>