MCTP: Address double free issue seen during istep 6.6
Hostboot was seeing intermittent crashes during the IPL where it
appeared that two threads were attempting to free the same pointer.
To avoid this we have added in a mutex around calls to the functions
that manipulate the tx queue in the openbmc's libmctp code. This
commit also removes the handle_inbound_messages thread instead handles
the KCS status immediately in the polling loop. This appeared to more
closely match what the BMC doing and it reduced the number of threads
potentially touching the MCTP data structures.

Change-Id: If4d2d6cb713cd31dbac8c85d3e27faf436856ed2
CQ: SW502741
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/102649
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>