Make host/SBE functional state mismatch errors unrecoverable
This commit makes mismatches between the ATTR_PG attribute Hostboot
calculates, and the functional state that the SBE calculates, cause
unrecoverable errors.

To make this work for standalone simics, we need to set SBE mailbox
scratch register #1 to the appropriate core mask based on the presence
information from MVPD. This is done in the standalone.simics script.

Change-Id: I9c5d520ef3d09c499db53effae1af5585f076e5d
RTC: 248496
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/102732
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>