DD2 ifCompiler support for processing DD2 scan initfiles
DD20 ifCompiler changes (specific to scan initfiles only):
Initfiles are now DD specific and are processed on a per DD level basis.
The dials in the initfile will only be attempted to be resolved against
engd/p10/<ddLevel> files. To do this, the initCompiler is envoked with
a single DD level. The above approach is different from P9/8, where
there was only one initfile (for each "chiplet domain" <cplt>) and which
would account for all possible DD levels via the use of DD specific
attribute conditionals in the spy expressions.

In p10, we now have two [input] initfiles for each <cplt>, e.g. cplt =
{core,iop,tp,..}. There is one initfile for each DD level as follows.
DD10 initfiles will keep their DD-less file names, ie p10.<cplt>.initfile
and referred to as p10.* in the following, whereas dedicated DD20
initfiles, if any, must be named p10.20.<cplt>.initfile and referred to
as p10.20.* in the following. If an p10.20.* initfile cannot be found,
then the corresponding p10.* initfile will be reused asis and copied to
p10.20.*.

As an alternative to providing a dedicated dd20 initfile described above,
it is allowed to put a dd20 conditional into the p10.* initfile so that
when the initfile is copied to p10.20.*, the dd20 conditional spy is
rejected during dd10 processing but accepted during dd20 processing.
This methodology may be more convenient if only a few spys need tweaking
for dd20.

Note that reusing the dd10 initfiles for dd20 comes with the added
responsibility of having to dd10 qualify spys that only are intended for
dd10 or which do not exist for dd20. As can be seen in the 104404 commit,
where we actually engage dd20, there's quite a bit of dd10 conditionals
that had to be added to existing dd10 initfiles to get it to work w/dd20
engd.

Additional DD levels:
The current implementation assumes only two DD levels, dd10 and dd20. It
is not coded in such a way that DD support can be expanded to three or
more DD levels just by updating chips.env.mk. If we decide to add dd30
the makefiles will have to manually be updated as well (not a
big deal).

Auto initfile discovery expansion:
The auto discovery of the dynamic input initfile [feature] names and xml
files has been updated to use the final renamed fully qualified initfile
names, as well as the [now seperately] copied <feature>.xml files, in the
/output/gen/initfiles/scan/<scan_folder> dir.

The same discovery process is now also used for the overlays initfiles.

It was not possible, in a practical way, to also do this for the base
rings. They still require manual attention. See scan_procedures.mk for
more info.

Target update to PCI initfiles:
In the original PCI/IOP initfiles, l_chip_ec and l_chip_id was not
declared in the generated HWP initfiles. The following changes have been
made to enable this declaration.
- ORing with TARGET_TYPE_PEC in Evaluator.C’s calculation of
  l_readChipEcAttrs,
- adding TARGET_TYPE_PROC_CHIP as 3rd target type in the PCI/IOP
  initfiles which enables a successful retrieval of chip_ec/id from the
  platform, and
- making corresponding updates to the PCI BASE target args in
  scan_procedures.mk.

All this is necessary just in case there should be an attribute qualifier
in the input initfile that requires the retrieval of chip_ec/id
attributes to properly apply the qualifier (and which would cause a call
to SpyLevelChipEcConditional which produces the qualifier conditionals).

Other changes:
- Added dependency on HWP initfiles production to the production
  of the HWP wrappers (this has been long overdue).
- Added dependency on the gathering of the scan inputs for the
  HWP wrappers (this also has been long overdue).
- The BUILD_*_INITFILE macro names have changed name to GET_*_SCAN_INPUT
  since that is a more accurate description of what those macros do.
- Some of the dynamic xml processing variable names have been renamed and
  removed to be consistent in how we name/use these (since there are so
  many of them.)

Key_Cronus_Test=XIP_REGRESS_SBE_QME

Change-Id: I8641e8b81902a1da1cd810db49e6d378a56b2c6c
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/102821
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Cronus HW CI <cronushw-ci+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-by: Claus Michael Olsen <cmolsen@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104358
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>