Remove File/Which.pm from Trace.pm
In an attempt to solve an issue where Trace.pm wastes minutes of time
trying to gather traces and then attempt to use fsp-trace without first
checking that it is usable, File/Which.pm was used as a quick way to fix
that issue. However, not all systems have that module installed. So,
that was removed in favor of an approach that will work on all linux
systems.

Change-Id: Iff35f632b9dc93107753e7ca292f94d5e153c430
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/102847
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>