Enable memory mirroring for P10
Up until this point in P10, memory mirroring has not been
enabled. With this commit is the FSP sets the regsitry
flags to enable mirrored memory hostboot will attempt
to load the reserved memory and payload in mirrored
address space. Note that this functionality is dependent
on another change which will move the HRMOR which hostboot
is being loaded at from 4GB - 256MB to be 4GB - 512 MB.
Without this hrmor change we will not have enough space
in the mirrored address window to load in the reserved
memory and payload.

RTC: 207396
Change-Id: I6b55c357a930225e29890ee8afa1f27ed5d58cc2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103466
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>