Add a virtual address range to handle large mallocs
As currently written, all malloc operations require contiguous
physical memory big enough to satisfy the requirements of the
caller.  This can be problematic if the memory becomes fragmented
into many pieces.  In that case we still have enough physical
memory to use, but it is scattered all around.

The changes here add the ability to request up to 20 allocations
of up to 1MB each.  The code allocates a virtual address range
that is then populated by arbitrary non-contiguous pages.  For
the external user the virtual memory is contiguous, but physically
the pages are scattered.

Change-Id: Ief82b4901b6f37f2fa5a2237f291deb33a4e546c
RTC: 257107
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103157
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>