p10_tod_setup: Increase I_Path delay range
Due to different frequencies and a better understanding for the
latency values involved, P10 TOD latency across two links exceeds
the previously allowed range of the I_Path delay register.

Fortunately we found two more register bits under the rug, so a
procedure change to allow larger values and set up a wider range
of bits is all that it takes.

Change-Id: I2497a3bfa97130f85db152fd41e4fb3ef27cf7c4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103479
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Reviewed-by: Douglas Holtsinger <douglas.holtsinger@ibm.com>
Reviewed-by: Jenny Huynh <jhuynh@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103482
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>