PRD: stale data used for OCMB attn workaround
After clearing OCMB attentions, and the associated bits in the
MC_DSTL_FIR, PRD must query the OCMB for any active attentions just in
case there were more than on attention on in the OCMB. If so, PRD will
set the associated bits in the MC_DSTL_FIR. Unfortunately, the current
workaround was using a cached value of the OCMB FIRs for the query.
Instead, PRD should re-read the value from hardware.

Change-Id: I9428e2a7bd9526f1af6bb855d915153b60bc389f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103595
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103608
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>