Fix bug in PLDM Fru IPZ data translation
Recently a change was made to have hostboot ignore PLDM Fru records
we do not know about. There was a bug in that fix that missed a
"is_ipz_record" check in the translation code. This was causing us
to get confused when we were passed a IPZ fru record we did not
expect and was causing a "invalid length" error later on when we
attempted to assemble the table of contents of the IPZ record we
were generating. This commit address this issue.

Change-Id: I4b97c7ff44a9b58f7c8c809949ed166386361608
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103600
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>