sbe: Enable L2 Loader w/ SBE FIFO
The L2 loader executes a sequence of scan, functional clock, scan to
load a cache image using the L2 store-queue when a functional fabric is
not present during cache-contained IPLs.

Each functional clock leads to a divergence in the ec_cl2_func rings
between cores in the same EQ. Subsequent multicast, parallel scans fail
with a compare error. There are several solutions:

	- Flush + scan the ec_cl2_func rings after every functional
	  clock (slow)
	- Perform all core scan + clock operations sequentially (slow)
	- Ignore the parallel-scan compare error (potentially masks
	  errors, but performant)

This patch implements the last option since the primary use of the
SBE_FIFO is to increase performance compared to using Cronus scans. The
check is only disabled during the initial "load" IPL phase of a
cache-contained IPL. All other IPLs are unaffected. For cache-contained
IPLs any parallel-scan compare error is still detected properly during
the contained-IPL phase later when parts of istep4 are re-executed.

Change-Id: Ib0bdd0af442490d30b89d6f84e69a5cfd23adce6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103713
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103724
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>