Dynamically find I2C Device Addresses for 4U DDIMM Devices
The main purpose of this commit is to use a new ekb function via
the FAPIWRAP interface to get the I2C Device Address for the 4U DDIMM
ADC and GPIO Expander devices.  It also makes some updates to address
review comments on the previouw 4U DDIMM targeting commit.

Change-Id: I6d7b534cdecdb67a660dd11df7108dfa632bfe21
RTC:214692
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103814
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>