Updated processMrw to inspect all bus connections associated with an IOHS
An IOHS target can have multiple bus connection with one being a connection
for SMP WRAP.  As the code was, it was only looking at the first bus connection
and was not always finding an SMP WRAP for the target.

This change corrects that.  This change inspects each bus connection, for said
target, looking to match the bus configuration the caller passed in via the '-c'
option.  If found, then the bus connection is made for the target and it's peer
target.

Change-Id: I92c2294d04d560814adff7ca4385c46ca9d1d100
RTC: 248352
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/103980
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>