EKB release-op920 build fix for RHEL7 & Copyright udpate
Cherry-pick commit 104130 in release-op930 to release-op920.
Updated various files/procedures as needed for RHEL7.

For UT-
Ignoring test result of platformcheck() in initfile_ci.py
as tests failing after cronus dev transitioned to
RHEL7/devtoolset-8 compile. The difference that is
triggering test to fail is not in ring data but in
ecmdDataBuffer output file footer. There is NO functional
diff in ring data. Since this test is being run on p9n
DD1 which is not useful as its not supported anymore,
wil be ignoring the test result as we can't remove
this test as follow-on tests continue to use the
env/ config file attribute setting setup by this test.

Change-Id: I76c7b5b129b0c4e8b1ba443a55f45a75c95f980a
CQ: SW508504
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104974
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: PPE CI <ppe-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: NAREN A DEVAIAH <naren.devaiah@in.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104980
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>