Fix OCMB firmware updates when OCMBFW PNOR part uses HPT
This commit fixes an issue in the OCMB firmware update code.

Previously, the code used the "secure protected payload size" of the
PNOR partition to determine the OCMB firmware size. However, when the
OCMBFW PNOR partition is built with a hashed page table (HPT), the
secure protected payload size is set to the size of the HPT itself,
instead of the size of the OCMB firmware blob, which makes the update
fail.

This commit fixes the issue by calculating the firmware size based on
whether the partition uses a HPT.

Change-Id: I67c8b1fb7a26191edf7c04dfa35b50a30928b83f
CQ: SW504971
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104446
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>