Pass complete input_args_t when reading records/keywords in ipvpd
In the retrieveRecord/retrieveKeyword function in ipvpd.C we were
only passing along the location member of input_args_t struct to
the fetchData function that performs the VPD read. When we do not
pass the entire input_args_t struct into fetchData, the fetchData
layer will assume that we want to AUTOSELECT where what EEPROM
value we are reading from ( CACHE vs HARDWARE ). This was breaking
the function cmpEecacheToEeprom because it was dropping off the
eepromSource member of the input_args_t struct that it was setting.

Change-Id: Ic7608cc104e535a57eab95f959dde63255ae69da
CQ: SW506013
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104466
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>