SBE Image Section Handling
Use proper API for determining if a section of an image exists.

The p9_xip_find API in the xipBinarySearch and xipLinearSearch
internal routines do not look in the Section Table for the
section names.

The p9_xip_get_section API will use the xipGetSectionPointer
to retrieve the Image section.  The size of the section requested
will return a size of zero for an empty section or the actual size
of the section requested.  With the proper query of the size of
the section we can determine whether to delete the section or
deterministically know that the section does -NOT- require deleting
(the delete puts added memory pressure during critical phases
of boot).

Early in p10 development it was determined that p9_xip_find would
not return proper data if the sections were present, therefore
we added defensive logic to basically always delete the sections
as a safeguard.

Due to the need to relieve memory pressure in critical paths
we needed to root cause why this exisitng code either had a bug
or was not being implemented properly.

Change-Id: I59ecc9034d6d651c6f7868f7747d35ca320cc456
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104505
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>