Adjustments to ATTR_PG/SBE scratch register enforcement
The SBE and Hostboot both independently calculate functional target
state. The SBE on the boot processor communicates the results of its
calculation to Hostboot via the SBE mailbox scratch registers, and
Hostboot verifies that it reaches the same conclusion to ensure that
we both operate on the same targets.

Historically, we checked on all processors in the host_gard step, and
we halted the IPL upon detecting a mismatch. The SBEs on the secondary
processors are not initialized by the time we reach host_gard, so a
previous commit narrowed the verification pass to operate only on the
boot processor.

This commit makes three more adjustments to the verification process:

1. The functional-state verification pass for the boot processor is
   moved to istep 7.5. This is the same place where we check the rest
   of the scratch registers for mismatches.

2. Instead of terminating the boot on a mismatch with the boot
   processor's SBE, we request a reconfig loop which will update the
   SBE and rectify the mismatch.

3. We perform the functional-state verification on the secondary
   processors in istep 8.4 where the SBEs are initialized. A mismatch
   in this step will terminate the IPL, because Hostboot itself pushes
   the values into the secondary processors' SBE scratch registers,
   which should never fail.

Change-Id: I92347f03183d3c7718288d0eb40e5f41a8efa692
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104523
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>