Fix addr_dis attribute setup for aggregate links
Procedure would disable coherent traffic on all valid links
before selecting the coherent link and allowing coherent
traffic on it. This was problematic since it would also
disable coherent traffic for non-aggregate links.

This commit reworks the procedure to disable coherent traffic
only on duplicate fabric links that have matching remote chip IDs
(that is, links that are within the set of aggregate links).

Change-Id: Ie5610d65109bec243ecf30522f1d1f8f6e53a382
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104654
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: Joseph J McGill <jmcgill@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104659
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>