Remove code to set HRMOR in secureboot signature
Previously there was some extra code in the script which generates
the HBB secureboot signature which was setting the HRMOR. This was
a holdover from the P8 design which required the HBB signature to
self describe where it was to be loaded. Now the SBE decides where
to load the HBB and set the HRMOR initially so this field in the
signature is irrelevant.

RTC: 207396
Change-Id: I30108cf15966ed40b4347a37cadd71aa07190d23
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/104719
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: William G Hoffa <wghoffa@us.ibm.com>